Register replacevars in the schema-blocks integration.
